LayoutTests/media/modern-media-controls/macos-fullscreen-media-controls/macos-fullscreen-media-controls-time-control-styles.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-utils.js" type="text/javascript"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

description("Testing the <code>MacOSFullscreenMediaControls</code> computed styles.");

window.jsTestIsAsync = true;

const mediaControls = new MacOSFullscreenMediaControls;

scheduler.frameDidFire = function()
{
    document.body.appendChild(mediaControls.element);

    shouldBeEqualToString("window.getComputedStyle(mediaControls.timeControl.element).left", "10px");
    shouldBeEqualToString("window.getComputedStyle(mediaControls.timeControl.element).top", "48px");
    shouldBeEqualToString("window.getComputedStyle(mediaControls.timeControl.element).width", "457px");
    shouldBeEqualToString("window.getComputedStyle(mediaControls.timeControl.scrubber.element).top", "-3px");
    shouldBeEqualToString("window.getComputedStyle(mediaControls.timeControl.elapsedTimeLabel.element).fontSize", "12px");
    shouldBeEqualToString("window.getComputedStyle(mediaControls.timeControl.remainingTimeLabel.element).fontSize", "12px");
    debug("");

    mediaControls.element.remove();

    finishMediaControlsTest();
};

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
